# :sparkles: gitfile: fetch no tags during clone